Cambios random para nuevos posts
- Se agregó SQLServer, por ahora no se usa SQLite
- Se usa Respawn para resetear tests
- Tabla AccessTokens para post nuevo

+++ tests/MediatRExample.IntegrationTests/TestBase.cs
@@ -1,10 +1,13 @@
 ﻿using MediatR;
+using MediatrExample.ApplicationCore.Domain;
 using MediatrExample.ApplicationCore.Features.Auth;
 using MediatrExample.ApplicationCore.Infrastructure.Persistence;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
+using Respawn;
 using System;
 using System.Linq.Expressions;
 using System.Net.Http;
@@ -24,9 +27,12 @@ public class TestBase
     public async Task<(HttpClient Client, string UserId)> CreateTestUser(string userName, string password, string[] roles)
     {
         using var scope = Application.Services.CreateScope();
-        var userManager = scope.ServiceProvider.GetRequiredService<UserManager<IdentityUser>>();
+        var userManager = scope.ServiceProvider.GetRequiredService<UserManager<User>>();
 
-        var newUser = new IdentityUser(userName);
+        var newUser = new User
+        {
+            UserName = userName
+        };
 
         await userManager.CreateAsync(newUser, password);
 
@@ -163,18 +169,34 @@ public class TestBase
 
         return result.AccessToken;
     }
-    /// <summary>
-    /// Se asegura de limpiar la BD
-    /// </summary>
-    /// <returns></returns>
+
+
     private async Task ResetState()
     {
         using var scope = Application.Services.CreateScope();
         var context = scope.ServiceProvider.GetService<MyAppDbContext>();
+        var userManager = scope.ServiceProvider.GetService<UserManager<User>>();
+        var roleManager = scope.ServiceProvider.GetService<RoleManager<IdentityRole>>();
 
-        context.Database.EnsureDeleted();
-        context.Database.EnsureCreated();
+        if (context.Database.IsSqlite())
+        {
+            context.Database.EnsureDeleted();
+            context.Database.EnsureCreated();
+        }
+        else if (context.Database.IsSqlServer())
+        {
+            var checkpoint = new Checkpoint
+            {
+                TablesToIgnore = new[] { "__EFMigrationsHistory" }
+            };
+            var config = scope.ServiceProvider.GetService<IConfiguration>();
+
+            await checkpoint.Reset(config.GetConnectionString("SqlServer"));
+
+        }
 
         await MyAppDbContextSeed.SeedDataAsync(context);
+        await MyAppDbContextSeed.SeedUsersAsync(userManager, roleManager);
     }
+
 }
